[tsk]三合一代码合并
