+++ JieLinkDevOpsApp/PartialView/PartialViewInterface/PartialViewInterface.csproj
@@ -133,6 +133,7 @@
     <Compile Include="Utils\NPOIExcelHelper.cs" />
     <Compile Include="Utils\ProcessHelper.cs" />
     <Compile Include="Utils\SerializationHelper.cs" />
+    <Compile Include="Utils\MSSqlHelper.cs" />
     <Compile Include="Utils\TaskHelper.cs" />
     <Compile Include="Utils\Win32API.cs" />
     <Compile Include="Utils\ZipHelper.cs" />
