+++ JieLinkDevOpsApp/PartialView/PartialViewOtherToJieLink/Models/TBaseHrRoomPos.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace PartialViewOtherToJieLink.Models
+{
+    public class TBaseHrRoomPos
+    {
+        public int ParsonId { get; set; }
+
+        public int MaxPos { get; set; }
+
+        public int UsedPos { get; set; }
+    }
+}
