+++ JieLinkDevOpsApp/PartialView/PartialViewOtherToJieLink/PartialViewOtherToJieLink.csproj
@@ -60,36 +60,43 @@
     <Compile Include="JSDSViewModels\TBaseEquipmentParamModel.cs" />
     <Compile Include="JSDSViewModels\TBaseEstablishModel.cs" />
     <Compile Include="JSDSViewModels\TBaseEstablishREquipmentModel.cs" />
-    <Compile Include="JSDSViewModels\TCacauthEstablishModel.cs" />
-    <Compile Include="JSDSViewModels\TCacauthServiceDistrictModel.cs" />
-    <Compile Include="Models\JsdsUpgradePolicy.cs" />
-    <Compile Include="Models\RecordComboxSelectType.cs" />
-    <Compile Include="ViewModels\ControlAccessPointGroup.cs" />
-    <Compile Include="ViewModels\ControlApDevcies.cs" />
-    <Compile Include="ViewModels\ControlAreaService.cs" />
-    <Compile Include="ViewModels\ControlDevices.cs" />
-    <Compile Include="ViewModels\ControlLeaseStall.cs" />
-    <Compile Include="ViewModels\ControlPerson.cs" />
-    <Compile Include="ViewModels\ControlRoleGroup.cs" />
-    <Compile Include="ViewModels\ControlVoucher.cs" />
-    <Compile Include="ViewModels\JsdsOneClickUpgradeViewModel.cs" />
-    <Compile Include="ViewModels\SysParameter.cs" />
-    <Compile Include="ViewModels\SystemSetDoorServiceToPerson.cs" />
     <Compile Include="JSDSViewModels\TBaseOrganizeModel.cs" />
     <Compile Include="JSDSViewModels\TBasePersonKeyModel.cs" />
     <Compile Include="JSDSViewModels\TBasePersonModel.cs" />
     <Compile Include="JSDSViewModels\TCacAccountModel.cs" />
+    <Compile Include="JSDSViewModels\TCacauthEstablishModel.cs" />
+    <Compile Include="JSDSViewModels\TCacauthServiceDistrictModel.cs" />
     <Compile Include="JSDSViewModels\TCacAuthServiceModel.cs" />
     <Compile Include="JSDSViewModels\TCacServiceSetModel.cs" />
+    <Compile Include="JSDSViewModels\TcacVoucherBizParamModel.cs" />
     <Compile Include="JSDSViewModels\TCacVoucherDataModel.cs" />
     <Compile Include="JSDSViewModels\TCacVoucherModel.cs" />
-    <Compile Include="JSDSViewModels\TcacVoucherBizParamModel.cs" />
     <Compile Include="JSDSViewModels\TParkPayModel.cs" />
     <Compile Include="JSDSViewModels\TParkpayrecordModel.cs" />
     <Compile Include="JSDSViewModels\TParkRecordInModel.cs" />
     <Compile Include="JSDSViewModels\TParkRecordOutModel.cs" />
     <Compile Include="JSDSViewModels\TSyncDictionary.cs" />
     <Compile Include="JSDSViewModels\TSysDictionary.cs" />
+    <Compile Include="Models\DbConnectionModel.cs" />
+    <Compile Include="Models\JsdsUpgradePolicy.cs" />
+    <Compile Include="Models\RecordComboxSelectType.cs" />
+    <Compile Include="Models\TBaseDeptModel.cs" />
+    <Compile Include="Models\TBaseHrPersonModel.cs" />
+    <Compile Include="Models\TBaseHrRoomPos.cs" />
+    <Compile Include="Models\TBaseJSRJCardInfoModel.cs" />
+    <Compile Include="Models\TBaseMcCardInfoModel.cs" />
+    <Compile Include="Models\TBaseMjDaKqCard.cs" />
+    <Compile Include="Models\TBaseTcCardInfoModel.cs" />
+    <Compile Include="Models\TBaseTcRecordModel.cs" />
+    <Compile Include="ViewModels\ControlLeaseStall.cs" />
+    <Compile Include="ViewModels\ControlPerson.cs" />
+    <Compile Include="ViewModels\ControlRoleGroup.cs" />
+    <Compile Include="ViewModels\ControlVoucher.cs" />
+    <Compile Include="ViewModels\JSDSToJieLinkViewModel.cs" />
+    <Compile Include="ViewModels\JSOCT2017ToJieLinkViewModel.cs" />
+    <Compile Include="ViewModels\JSRJ1116ToJieLinkViewModel.cs" />
+    <Compile Include="ViewModels\SysParameter.cs" />
+    <Compile Include="ViewModels\SystemSetDoorServiceToPerson.cs" />
     <Page Include="JsdsOneClickUpgradeToJieLink.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
